git-svn-id: https://svn.code.sf.net/p/clinkc/code/trunk@401 bfd3be0d-9f21-0410-925f-e30e9767a5cc

+++ wrapper/objc/CyberLink/CGUpnpAction.h
@@ -68,5 +68,5 @@ typedef void CgUpnpAction;
  *
  * @return The status code
  */
-- (int)statusCode;
+- (NSInteger)statusCode;
 @end
